requirement.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
langchain
langchain_google_genai
PyPDF2
